requirements.txt
anthropic
os
Pillow
PyMuPDF
openai==0.28
python-dotenv
base64
requests
io
datetime
